+++ lib/td_auth_web/views/user_view.ex
@@ -32,6 +32,14 @@ defmodule TdAuthWeb.UserView do
     Map.take(user, [:id, :user_name, :email, :full_name, :role])
   end
 
+  def render("search.json", %{users: users}) do
+    %{data: render_many(users, UserView, "user_basic.json")}
+  end
+
+  def render("user_basic.json", %{user: user}) do
+    Map.take(user, [:id, :email, :full_name])
+  end
+
   defp render_acls(user_map, %{acls: acls}) do
     Map.put(user_map, :acls, acls)
   end